fix props docgen
